requirements.txt
# Orchestration & LLMs
crewai
crewai[tools]
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-mcp-adapters
google-generativeai

# MCP & Web Scraping
mcp
playwright
httpx

# Data Processing & Vector DB
pandas
openpyxl
qdrant-client
sentence-transformers
pypdf
Pillow
fpdf

# Backend API
fastapi
uvicorn[standard]
apscheduler
python-dotenv
